// projects/mercado-ion-mobile/src/app/shared/types/models/User.ts
export interface IUser {
  user: any;
  userObj: any;
  id?: string;
  username: any;
  email: string;
  provider?: string;
  confirmed?: boolean;
  blocked?: boolean;
  role?: any;
  created_by?: any;
  updated_by?: any;
  avatar?: any;
}